Add Tabs to subscriptions page for live streams and shorts
