+++ src/renderer/views/Subscriptions/Subscriptions.css
@@ -8,20 +8,31 @@
   color: var(--tertiary-text-color);
 }
 
-.floatingTopButton {
-  position: fixed;
-  top: 70px;
-  right: 10px;
+.subscriptionTabs {
+  width: 100%;
+  margin-top: -3px;
+  color: var(--tertiary-text-color);
+  margin-bottom: 10px;
 }
 
-.channelBubble {
-  display: inline-block;
+.selectedTab {
+  border-bottom: 3px solid var(--primary-color);
+  color: var(--primary-text-color);
+  font-weight: bold;
+  box-sizing: border-box;
+  margin-bottom: -3px;
 }
 
-@media only screen and (max-width: 350px) {
-  .floatingTopButton {
-    position: absolute
-  }
+.tab {
+  text-align: center;
+  padding: 15px;
+  font-size: 15px;
+  cursor: pointer;
+  align-self: flex-end;
+}
+
+.tab:hover {
+  font-weight: bold;
 }
 
 @media only screen and (max-width: 680px) {
